Prepare release with support reviews and deployment config

+++ backend/src/Domain/Entities/Overtime.cs
@@ -5,7 +5,7 @@ public class Overtime
     public Guid Id { get; set; }
     public Guid ShiftScheduleId { get; set; }
     public DateTime OvertimeDate { get; set; }
-    public int Hours { get; set; } // Количество часов переработки
+    public decimal Hours { get; set; } // Количество часов переработки в формате HH.MM
     public decimal Amount { get; set; } // Рассчитанная сумма для этой переработки
     public string? Notes { get; set; } // Дополнительные заметки
