+++ backend/src/Domain/Entities/ShiftSchedule.cs
@@ -8,7 +8,7 @@ public class ShiftSchedule
     public DateTime StartDate { get; set; }
     public int WorkDays { get; set; }
     public int OffDays { get; set; }
-    public int HoursPerShift { get; set; }
+    public decimal HoursPerShift { get; set; }
     public int Months { get; set; }
     public bool NightFirst { get; set; }
     public string CalendarJson { get; set; } = string.Empty;
